solana-ai-trading-bot/backend/requirements.txt
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
websockets==12.0
sqlalchemy==2.0.30
asyncio==3.4.3
python-jose[cryptography]==3.3.0
pycryptodome==3.20.0
pysolana==0.1.1
base58==2.1.1
PyNaCl==1.5.0
loguru==0.7.2
python-multipart==0.0.9
fastapi==0.104.1
solana==0.30.0
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
cryptography==41.0.4
psutil==5.9.6
google-generativeai==0.3.2
aiohttp